Feat: Refactor data model and update references for consistency
Replaced outdated and redundant fields (e.g., `reference`) and migrated numeric fields (e.g., `monthCost`, `salaryByHour`) to `BigDecimal` for improved precision. Updated queries, entities, and DTOs to align with the new structure, ensuring seamless integration and consistency across the application.

+++ src/main/java/com/edusyspro/api/dto/ClasseDTO.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.List;
 
@@ -29,7 +30,7 @@ public class ClasseDTO {
     private List<EnrollmentDTO> students;
     @JsonProperty("classeTeachers")
     private List<TeacherDTO> classTeacherCourses;
-    private Double monthCost;
+    private BigDecimal monthCost;
     private ZonedDateTime createdAt;
     private ZonedDateTime modifyAt;
