+++ src/main/java/com/edusyspro/api/repository/ClasseRepository.java
@@ -11,6 +11,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -67,7 +68,7 @@ public interface ClasseRepository extends JpaRepository<ClasseEntity, Integer> {
             String classeCategory,
             int gradeId,
             int roomNumber,
-            double monthCost,
+            BigDecimal monthCost,
             int classeId
     );
